add module handlerAddComments
