+++ variables.js
@@ -4,6 +4,7 @@ export const loading = document.querySelector('.load');
 export const inputComment = form.querySelector('.add-form-text');
 export const inputName = form.querySelector('.add-form-name');
 export const formButton = form.querySelector('.add-form-button');
-
+// export let inputNameValue = '';
+// export let inputCommentValue = '';
